Fix fonts in iOS 13
Apple does not allow explicit use of the various
system fonts. To get around this, I was originally
init'ing the fonts I wanted to use then taking
their `fontName`s for use later where I wanted
them. In iOS 13, this began returning Times New
Roman for everything (gross).

Not sure if this is a bug or a deliberate new
obfuscation, but what I was doing was admittedly
a bit hacky anyway. So now I'm using the original
constructed font each time I need it and simply
updating its size.

Still not necessarily safe from Apple's
implementation, they could eventually decide to
change the used system font on the fly when its
size changes, but it works. Hopefully it's a bit
safer this time.

But also this app could be dead in a couple weeks
anyway so #yolo.