Fix PlayerSearchVC retain cycle
This was happening because the method was being
passed to a view's transform function, which then
passed it into an observer of the binding also
stored on the view controller.

This is my own dumb fault and a symptom of how
close to the line BindableViews already walks with
having to be overly aware of what you're doing.
It's not good.

This could be combatted by a view controller
removing all observers on all its state when it
is removed from its window, but that's an
unreasonable ask for every controller, and
building it into the super class proved impossible
due to the genericism of State.

Really, BindableViews should take the stance of
SwiftUI and not allow arbitrary observations of
its Bindings. They should be reserved exclusively
for views and views should be the only responders
of changes to those bindings.

That'll be an overhaul and will likely coincide
with moving State and Binding to @propertyWrappers
now that Swift 5.1 is out of beta. A fun project,
but not for today.