Adjust Glory Profile layout for x320 screens
Scale down some font sizes and spacing values

Note: on iPhone SE iOS 10.3.1 we get none of the
specific system fonts we want inside of the cards.
There doesn't appear to be any good way around
this. It looks bad, but it is an edge case. Fonts
are (mostly) correct on iPhone SE iOS 12.