Show more errors on player search
Was previously only showing bungie maintenance
errors, but this was insufficient as it's possible
for the API to be down but not in maintenance mode
(as we saw yesterday). Also, this will allow users
to see when e.g. their network connection is
preventing searches from completing.

My original intention for not showing errors was
to avoid showing messages every single time a user
inputs a bad name as its intuitive that a search
which doesn't result in the button being enabled
did not return results; a message is unneeded.

But really I should have always checked specifically
for the errors we want to avoid and let others
through. So my b.